test(skore): Move comparison regression fixture to conftest
