+++ skore/tests/unit/displays/prediction_error/test_comparison_estimator.py
@@ -324,37 +324,11 @@ def test_title(pyplot, linear_regression_with_train_test):
     assert "Data source: Test set" in title
 
 
-@pytest.fixture
-def linear_regression_comparison_report(linear_regression_with_train_test):
-    """Fixture providing a ComparisonReport with two linear regression estimators."""
-    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
-    estimator_2 = clone(estimator).fit(X_train, y_train)
-    report = ComparisonReport(
-        reports={
-            "estimator_1": EstimatorReport(
-                estimator,
-                X_train=X_train,
-                y_train=y_train,
-                X_test=X_test,
-                y_test=y_test,
-            ),
-            "estimator_2": EstimatorReport(
-                estimator_2,
-                X_train=X_train,
-                y_train=y_train,
-                X_test=X_test,
-                y_test=y_test,
-            ),
-        }
-    )
-    return report
-
-
 def test_regression_data_source_both(pyplot, linear_regression_comparison_report):
     """Regression test: `data_source='both'` should plot without crashing."""
-
     report = linear_regression_comparison_report
     display = report.metrics.prediction_error(data_source="both")
+
     assert isinstance(display, PredictionErrorDisplay)
 
     display.plot()
